perf(translator): .length i64 flavor; refresh lev bench + matrix
flavor.rs: expr_flavor + structural_flavor add a `.length` -> I64 arm
(string/array length is a non-negative integer < 2^53), in lockstep with
the `as i64` emit. member.rs: the four `.length` emit sites switch
`as f64` -> `as i64` (static member, optional field, optional chain,
DsMatch captures). transpile_js.rs: the homogeneous-array test now expects
`xs.len() as i64`.

levenshtein: n/m (from a.length/b.length) + sc + lst + nm1 stay in i64
instead of round-tripping through f64 per op. 78.6 -> 59.9 ms (1.30x); ds
vs node 2.16x / vs bun 2.05x. lib 798 passed.

peq Vec<i32> PoC ruled out: the translator-faithful version (array_set
generic i64 + eq i64, keeping the ES ToInt32 hop and the auto-grow check)
is only 1.10x (66.6 -> 60.4 ms) -- auto-grow overhead cancels the element
cast savings, and a global array-element flavor system is not worth it for
a single bench (KISS/YAGNI). charCodeAt ascii direct read breaks the ES
out-of-bounds NaN semantics (no range analysis) and is out too.

bench: refresh the levenshtein row (ds/node/bun, 9 samples).

matrix: refresh the test262 string/array baseline. The large delta (string
667->450, array 2622->1169 supported) is the four-state conformance
methodology strictification -- engine-fallback build failures are now
`partial`, not `supported` -- not a `.length` i64 regression. The one
i64/f64 E0308 in reduce 15.4.4.21-9-c-ii-17 is `preIteration = 1` Phase 1
integer-literal flavor (the fixture has no `.length`).

+++ crates/dashscript/src/translator/expressions/member.rs
@@ -28,10 +28,10 @@ pub(super) fn chain_expr(elem: &oxc_ast::ast::ChainElement, ctx: &Ctx<'_>) -> Ex
             let field = bindings::snake(&sm.property.name);
             // `.length` on a Vec/String maps to Rust's `.len()` (a method, not a
             // field), mirroring the non-chain path; `len()` returns `usize` so
-            // cast to `f64` (TS `.length` is always a `number`).
+            // cast to `i64` (TS `.length` is an integer < 2^53 — see `flavor.rs`).
             if field == "length" {
                 let c = Ident::new("__c", Span::call_site());
-                return parse_quote!(#obj.as_ref().map(|#c| (#c.len() as f64)));
+                return parse_quote!(#obj.as_ref().map(|#c| (#c.len() as i64)));
             }
             if receiver_field_is_optional(&sm.object, &field, ctx) {
                 parse_quote!(#obj.as_ref().and_then(|__c| __c.#field.clone()))
@@ -217,16 +217,17 @@ pub(super) fn member_expr(sm: &StaticMemberExpression, ctx: &Ctx<'_>) -> Expr {
     }
     // `m.index`/`m.input`/`m.length` on a `let m = s.match(/pat/)` result → the
     // `DsMatch` fields. Checked before the generic `.length` arm, which would
-    // try `Option<DsMatch>::len()` (no such method). `index`/`length` are ES
-    // numbers (cast to `f64`); `input` is the haystack string. (A bare
-    // `m.groups` — the whole groups object — is not yet handled; only the
-    // `.groups.name` access above.)
+    // try `Option<DsMatch>::len()` (no such method). `index` is an ES number
+    // (cast to `f64`); `length` is an integer < 2^53 (cast to `i64`, matching
+    // `flavor.rs`); `input` is the haystack string. (A bare `m.groups` — the
+    // whole groups object — is not yet handled; only the `.groups.name` access
+    // above.)
     if is_match_local(&sm.object, ctx) {
         let obj = translate_expr(&sm.object, ctx);
         match field_name {
             "index" => return parse_quote!(#obj.as_ref().unwrap().index as f64),
             "input" => return parse_quote!(#obj.as_ref().unwrap().input.clone()),
-            "length" => return parse_quote!(#obj.as_ref().unwrap().captures.len() as f64),
+            "length" => return parse_quote!(#obj.as_ref().unwrap().captures.len() as i64),
             _ => {}
         }
     }
@@ -440,8 +441,10 @@ pub(super) fn member_expr(sm: &StaticMemberExpression, ctx: &Ctx<'_>) -> Expr {
     let obj =
         option_unwrap_object(&sm.object, ctx).unwrap_or_else(|| translate_expr(&sm.object, ctx));
     // `.length` on a Vec/String maps to Rust's `.len()` (a method, not a field).
-    // TS `.length` is always a `number` → `f64`; `len()` returns `usize`, so cast.
-    // Index/repeat sites that need `usize` cast the whole expression again.
+    // TS `.length` is a non-negative integer < 2^53 → `i64` (see `flavor.rs`);
+    // `len()` returns `usize`, so cast. A binding fed `.length` then keeps the
+    // integer flavor, skipping the `f64` cast chain in tight loops. Index/repeat
+    // sites that need `usize` cast the whole expression again.
     if field_name == "length" {
         // `.length` on an optional Vec field (`parent.elements.length`) — the
         // field is `Option<Vec<..>>`, and `Option::len()` is private (nightly);
@@ -452,11 +455,11 @@ pub(super) fn member_expr(sm: &StaticMemberExpression, ctx: &Ctx<'_>) -> Expr {
             if static_member_is_optional_field(&inner.object, &inner_field, ctx) {
                 let inner_obj = translate_expr(&inner.object, ctx);
                 return parse_quote!(
-                    (#inner_obj.#inner_field.as_ref().map(|__c| __c.len() as f64).unwrap_or(0_f64))
+                    (#inner_obj.#inner_field.as_ref().map(|__c| __c.len() as i64).unwrap_or(0_i64))
                 );
             }
         }
-        return parse_quote!((#obj.len() as f64));
+        return parse_quote!((#obj.len() as i64));
     }
     let field = bindings::snake(field_name);
     parse_quote!(#obj.#field)
